+++ src/app.module.ts
@@ -9,7 +9,10 @@ import { CategoriesModule } from './categories/categories.module';
 import { CommentsModule } from './comments/comments.module';
 import { authConfig } from './config/auth.config';
 import { commentReportConfig } from './config/comment-report.config';
+import { geocodingConfig } from './config/geocoding.config';
 import { offerExpirationConfig } from './config/offer-expiration.config';
+import { GeocodingModule } from './geocoding/geocoding.module';
+import { MerchantsModule } from './merchants/merchants.module';
 import { ModerationModule } from './moderation/moderation.module';
 import { OffersModule } from './offers/offers.module';
 import { PrismaModule } from './prisma/prisma.module';
@@ -27,6 +30,7 @@ import { VotesModule } from './votes/votes.module';
         reportConfig,
         offerExpirationConfig,
         commentReportConfig,
+        geocodingConfig,
       ],
     }),
     ScheduleModule.forRoot(),
@@ -43,6 +47,8 @@ import { VotesModule } from './votes/votes.module';
     ModerationModule,
     CommentsModule,
     CategoriesModule,
+    MerchantsModule,
+    GeocodingModule,
   ],
   controllers: [AppController],
   providers: [