+++ src/offers/offers.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 
+import { MerchantsModule } from '../merchants/merchants.module';
 import { PrismaModule } from '../prisma/prisma.module';
 import { OfferOwnerGuard } from './guards/offer-owner.guard';
 import { OffersController } from './offers.controller';
@@ -7,7 +8,7 @@ import { OffersExpirationService } from './offers-expiration.service';
 import { OffersService } from './offers.service';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, MerchantsModule],
   controllers: [OffersController],
   providers: [OffersService, OfferOwnerGuard, OffersExpirationService],
   exports: [OffersService, OfferOwnerGuard, OffersExpirationService],